requirements.txt
python-dateutil==2.7.5
openapi-specgen==0.0.3dev2
